varias validaciones

+++ base/funciones/kaf.ft_movimiento_af_ime.sql
@@ -1,8 +1,13 @@
-CREATE OR REPLACE FUNCTION "kaf"."ft_movimiento_af_ime" (	
-				p_administrador integer, p_id_usuario integer, p_tabla character varying, p_transaccion character varying)
-RETURNS character varying AS
-$BODY$
+--------------- SQL ---------------
 
+CREATE OR REPLACE FUNCTION kaf.ft_movimiento_af_ime (
+  p_administrador integer,
+  p_id_usuario integer,
+  p_tabla varchar,
+  p_transaccion varchar
+)
+RETURNS varchar AS
+$body$
 /**************************************************************************
  SISTEMA:		Sistema de Activos Fijos
  FUNCION: 		kaf.ft_movimiento_af_ime
@@ -28,6 +33,7 @@ DECLARE
 	v_mensaje_error         text;
 	v_id_movimiento_af		integer;
 	v_id_cat_estado_fun		integer;
+    v_registros				record;
 			    
 BEGIN
 
@@ -36,14 +42,26 @@ BEGIN
 
 	/*********************************    
  	#TRANSACCION:  'SKA_MOVAF_INS'
- 	#DESCRIPCION:	Insercion de registros
- 	#AUTOR:		admin	
- 	#FECHA:		18-03-2016 05:34:15
+ 	#DESCRIPCION:	Insercion de registros, validaciones de estado y cantidad
+ 	#AUTOR:		admin ,RAC	
+ 	#FECHA:		18-03-2016, 23/03/2017
 	***********************************/
 
 	if(p_transaccion='SKA_MOVAF_INS')then
 					
         begin
+        
+            select
+              mov.estado,
+              mov.codigo
+             INTO 
+              v_registros
+            from kaf.tmovimiento mov
+            where mov.id_movimiento = v_parametros.id_movimiento;
+            
+            IF v_registros.estado != 'borrador' THEN
+               raise exception 'Solo puede insertar acctivos en movimientos en borrador';
+            END IF;
 
         	--Obtiene estado funcional del activo fijo
         	select
@@ -52,36 +70,48 @@ BEGIN
         	v_id_cat_estado_fun
         	from kaf.tactivo_fijo
         	where id_activo_fijo = v_parametros.id_activo_fijo;
+            
+            --verificamos que el activo no este duplicado
+            
+            IF EXISTS(SELECT 1 
+                      from kaf.tmovimiento_af maf 
+                      where     maf.id_movimiento =  v_parametros.id_movimiento 
+                           and  maf.id_activo_fijo = v_parametros.id_activo_fijo 
+                           and maf.estado_reg = 'activo') THEN
+                 raise exception 'El activo ya se encuentre registro para este movimiento';
+            END IF;     
+            
+            
 
         	--Sentencia de la insercion
         	insert into kaf.tmovimiento_af(
-			id_movimiento,
-			id_activo_fijo,
-			id_cat_estado_fun,
-			id_movimiento_motivo,
-			estado_reg,
-			importe,
-			vida_util,
-			fecha_reg,
-			usuario_ai,
-			id_usuario_reg,
-			id_usuario_ai,
-			id_usuario_mod,
-			fecha_mod
+                id_movimiento,
+                id_activo_fijo,
+                id_cat_estado_fun,
+                id_movimiento_motivo,
+                estado_reg,
+                importe,
+                vida_util,
+                fecha_reg,
+                usuario_ai,
+                id_usuario_reg,
+                id_usuario_ai,
+                id_usuario_mod,
+                fecha_mod
           	) values(
-			v_parametros.id_movimiento,
-			v_parametros.id_activo_fijo,
-			v_id_cat_estado_fun,
-			v_parametros.id_movimiento_motivo,
-			'activo',
-			v_parametros.importe,
-			v_parametros.vida_util,
-			now(),
-			v_parametros._nombre_usuario_ai,
-			p_id_usuario,
-			v_parametros._id_usuario_ai,
-			null,
-			null
+                v_parametros.id_movimiento,
+                v_parametros.id_activo_fijo,
+                v_id_cat_estado_fun,
+                v_parametros.id_movimiento_motivo,
+                'activo',
+                v_parametros.importe,
+                v_parametros.vida_util,
+                now(),
+                v_parametros._nombre_usuario_ai,
+                p_id_usuario,
+                v_parametros._id_usuario_ai,
+                null,
+                null
 			)RETURNING id_movimiento_af into v_id_movimiento_af;
 			
 			--Definicion de la respuesta
@@ -103,26 +133,50 @@ BEGIN
 	elsif(p_transaccion='SKA_MOVAF_MOD')then
 
 		begin
-			--Obtiene estado funcional del activo fijo
+			
+            select
+              mov.estado,
+              mov.codigo
+             INTO 
+              v_registros
+            from kaf.tmovimiento mov
+            where mov.id_movimiento = v_parametros.id_movimiento;
+            
+            IF v_registros.estado != 'borrador' THEN
+               raise exception 'Solo puede modificar acctivos en movimientos en borrador';
+            END IF;
+            
+            
+            --Obtiene estado funcional del activo fijo
         	select
         	id_cat_estado_fun
         	into
         	v_id_cat_estado_fun
         	from kaf.tactivo_fijo
         	where id_activo_fijo = v_parametros.id_activo_fijo;
+            
+           --  verificamos que el activo no este duplicado
+            
+            IF EXISTS(SELECT 1 
+                      from kaf.tmovimiento_af maf 
+                      where     maf.id_movimiento =  v_parametros.id_movimiento 
+                           and  maf.id_activo_fijo = v_parametros.id_activo_fijo 
+                           and maf.estado_reg = 'activo'  and maf.id_movimiento_af != v_parametros.id_movimiento_af) THEN
+                 raise exception 'El activo ya se encuentre registro para este movimiento';
+            END IF; 
 
 			--Sentencia de la modificacion
 			update kaf.tmovimiento_af set
-			id_movimiento = v_parametros.id_movimiento,
-			id_activo_fijo = v_parametros.id_activo_fijo,
-			id_cat_estado_fun = v_id_cat_estado_fun,
-			id_movimiento_motivo = v_parametros.id_movimiento_motivo,
-			importe = v_parametros.importe,
-			vida_util = v_parametros.vida_util,
-			id_usuario_mod = p_id_usuario,
-			fecha_mod = now(),
-			id_usuario_ai = v_parametros._id_usuario_ai,
-			usuario_ai = v_parametros._nombre_usuario_ai
+                id_movimiento = v_parametros.id_movimiento,
+                id_activo_fijo = v_parametros.id_activo_fijo,
+                id_cat_estado_fun = v_id_cat_estado_fun,
+                id_movimiento_motivo = v_parametros.id_movimiento_motivo,
+                importe = v_parametros.importe,
+                vida_util = v_parametros.vida_util,
+                id_usuario_mod = p_id_usuario,
+                fecha_mod = now(),
+                id_usuario_ai = v_parametros._id_usuario_ai,
+                usuario_ai = v_parametros._nombre_usuario_ai
 			where id_movimiento_af=v_parametros.id_movimiento_af;
                
 			--Definicion de la respuesta
@@ -144,6 +198,19 @@ BEGIN
 	elsif(p_transaccion='SKA_MOVAF_ELI')then
 
 		begin
+        
+            select
+              mov.estado,
+              mov.codigo
+             INTO 
+              v_registros
+            from kaf.tmovimiento mov
+            inner join kaf.tmovimiento_af maf on maf.id_movimiento = mov.id_movimiento
+            where maf.id_movimiento_af = v_parametros.id_movimiento_af;
+            
+            IF v_registros.estado != 'borrador' THEN
+               raise exception 'Solo puede retirar activos en movimientos en borrador';
+            END IF;
 			--Sentencia de la eliminacion
 			delete from kaf.tmovimiento_af
             where id_movimiento_af=v_parametros.id_movimiento_af;
@@ -173,7 +240,9 @@ EXCEPTION
 		raise exception '%',v_resp;
 				        
 END;
-$BODY$
-LANGUAGE 'plpgsql' VOLATILE
+$body$
+LANGUAGE 'plpgsql'
+VOLATILE
+CALLED ON NULL INPUT
+SECURITY INVOKER
 COST 100;
-ALTER FUNCTION "kaf"."ft_movimiento_af_ime"(integer, integer, character varying, character varying) OWNER TO postgres;
